--- Clasroom/src/main/java/com/minsait/Clasroom/exceptions/GlobalExceptionHandler.java
@@ -1,23 +1,0 @@
-package com.minsait.Clasroom.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<?> sqlExceptionHanlder(SQLException exception){
-        Map<String, Object> response = new HashMap<>();
-        response.put("status", exception.getErrorCode());
-        response.put("message", exception.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-}
