--- Clasroom/src/main/java/com/minsait/Clasroom/repositories/ClassroomRepository.java
@@ -1,7 +1,0 @@
-package com.minsait.Clasroom.repositories;
-
-import com.minsait.Clasroom.models.entities.Classroom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
-}
