--- Clasroom/src/main/resources/application.yml
@@ -1,15 +1,0 @@
-spring:
-  application:
-    name: msvc-classroom
-  datasource:
-    url: jdbc:postgresql://localhost:5432/SchoolSystem
-    username: postgres
-    password: root
-  config:
-    import: optional:configserver:http://localhost:8888
-  jpa:
-    show:
-      sql: true
-    properties:
-      hibernate:
-        format_sql: true
