--- Clasroom/src/test/java/com/minsait/Clasroom/ClasroomApplicationTests.java
@@ -1,13 +1,0 @@
-package com.minsait.Clasroom;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ClasroomApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
