+++ Classroom/src/main/java/com/minsait/Classroom/ClasroomApplication.java
@@ -1,0 +1,15 @@
+package com.minsait.Classroom;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableDiscoveryClient
+@EnableFeignClients
+@SpringBootApplication
+public class ClasroomApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(ClasroomApplication.class, args);
+	}
+}
