+++ Classroom/src/main/java/com/minsait/Classroom/models/dto/StudentProfileDTO.java
@@ -1,0 +1,15 @@
+package com.minsait.Classroom.models.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter @Setter
+public class StudentProfileDTO {
+
+    public String first_name;
+    public String last_name;
+    public String email;
+
+}
