+++ Classroom/src/main/java/com/minsait/Classroom/models/dto/SubjectDTO.java
@@ -1,0 +1,4 @@
+package com.minsait.Classroom.models.dto;
+
+public class SubjectDTO {
+}
