+++ Classroom/src/main/java/com/minsait/Classroom/repositories/ClassroomRepository.java
@@ -1,0 +1,7 @@
+package com.minsait.Classroom.repositories;
+
+import com.minsait.Classroom.models.entities.Classroom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
+}
