+++ Classroom/src/main/resources/application.yml
@@ -1,5 +1,3 @@
-server:
-  port: 8084
 spring:
   application:
     name: msvc-classroom
@@ -7,12 +5,11 @@ spring:
     url: jdbc:postgresql://localhost:5432/prueba
     username: postgres
     password: root
+  config:
+    import: optional:configserver:http://localhost:8888
   jpa:
+    show:
+      sql: true
     properties:
       hibernate:
-        jdbc:
-          lob:
-            non_contextual_creation: true
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-    hibernate:
-      ddl-auto: create-drop
+        format_sql: true
