+++ Classroom/src/test/java/com/minsait/Classroom/ClasroomApplicationTests.java
@@ -1,0 +1,13 @@
+package com.minsait.Classroom;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ClasroomApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
