+++ Configuration/src/main/resources/configurations/msvc-students.yml
@@ -11,9 +11,9 @@ spring:
       ddl-auto: create-drop
       show-sql: true
   datasource:
-    url: jdbc:postgresql://localhost:5432/SchoolSystem
+    url: jdbc:postgresql://localhost:5432/prueba
     username: postgres
-    password: postgres
+    password: root
 
 springdoc:
   swagger-ui.path: /swagger-ui.html
