+++ Configuration/src/main/resources/configurations/msvc-teacher.yml
@@ -4,9 +4,9 @@ spring:
   application:
     name: msvc-teacher
   datasource:
-    url: jdbc:postgresql://localhost:5432/SchoolSystem
+    url: jdbc:postgresql://localhost:5432/prueba
     username: postgres
-    password: postgres
+    password: root
     driver-class-name: org.postgresql.Driver
   jpa:
     properties:
