+++ Students/src/main/resources/application.yml
@@ -12,6 +12,6 @@ spring:
       ddl-auto: update
       show-sql: true
   datasource:
-    url: jdbc:postgresql://localhost:5432/SchoolSystem
+    url: jdbc:postgresql://localhost:5432/prueba
     username: postgres
-    password: postgres
+    password: root
