+++ pom.xml
@@ -32,7 +32,7 @@
         <module>Students</module>
         <module>Teacher</module>
         <module>Subject</module>
-        <module>Clasroom</module>
+        <module>Classroom</module>
     </modules>
 
    <dependencies>
